requirements.txt
networkx>=2.6.0
matplotlib>=3.4.0
watchdog>=2.1.0
pyvis>=0.1.9
# pygraphviz>=1.7.0  # Optional, for hierarchical layouts
typing-extensions>=4.0.0  # For improved type hints
pymongo>=4.3.0  # For MongoDB operations
# sentence-transformers>=2.2.0  # For generating embeddings - commented out
# pinecone-client>=2.2.0  # For vector database operations - commented out
python-dotenv>=0.21.0  # For environment variables
